src/mock-api/requirements.txt
Flask==2.3.2
Flask-HTTPAuth==4.8.0
pendulum==2.1.2
python-dateutil==2.8.2
pytz==2023.3
pytzdata==2020.1
requests==2.31.0
